pyproject.toml
[tool.poetry]
name = "espritgames-payment"
version = "0.1.0"
description = ""
authors = ["Klikovskiy <biisky.k@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.3"
python-dotenv = "^1.0.1"
djangorestframework = "^3.15.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
